CORE-2534 Added group references to AnalysisListing.
Added a list of AnalysisGroup model references to AnalysisListing.
Updated AnalysisGroup::filterAllAnalysesByNameOrDesc method to add the group to each of its analyses in the filtered results.
This will allow metadactyl to return the visible group IDs the analyses is listed in for each analysis JSON result in the service response.